refactor(quartz): extract sync logic into IQuartzScheduleSynchronizer
Background worker semantics in .NET 10 changed (ExecuteAsync now runs
entirely on a background thread, so IHost.StartAsync returns before any
of it runs). Per project guidance, background workers should be thin
shells and business logic should live in services so it can be tested
directly.

Extracts SyncSchedulesAsync + helpers out of QuartzSchedulingSyncService
into a new IQuartzScheduleSynchronizer. Worker becomes a thin orchestration
loop. Integration tests exercise the service directly with a real RAM
Quartz scheduler — no IHost, no SetSchedulerForTesting, no
[InternalsVisibleTo].

Drops test-only hooks added in the prior commit:
- SetSchedulerForTesting method (deleted)
- internal visibility on SyncSchedulesAsync (moot — method deleted)
- TelegramGroupsAdmin.BackgroundJobs/AssemblyInfo.cs (deleted)

#459 fix preserved: orphan-removal still sources its registered-names set
from BackgroundJobNames.AllRegisteredNames, now inside the new service.
