--- TelegramGroupsAdmin.BackgroundJobs/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("TelegramGroupsAdmin.IntegrationTests")]
